// src/types/rs-types/rs-svartype.ts
export enum RSSvartype {
    JA_NEI = 'JA_NEI',
    CHECKBOX = 'CHECKBOX',
    CHECKBOX_GRUPPE = 'CHECKBOX_GRUPPE',
    CHECKBOX_PANEL = 'CHECKBOX_PANEL',
    DATOER = 'DATOER',
    BELOP = 'BELOP',
    KILOMETER = 'KILOMETER',
    KVITTERING = 'KVITTERING',
}

export type RSSvartypeType =
    'JA_NEI' |
    'CHECKBOX' |
    'CHECKBOX_GRUPPE' |
    'CHECKBOX_PANEL' |
    'DATOER' |
    'BELOP' |
    'KILOMETER' |
    'KVITTERING'